+++ modules/test-runner/package.json
@@ -7,7 +7,8 @@
     "build": "webpack --config ops/webpack.config.js",
     "lint": "../../node_modules/.bin/eslint -c '../../.eslintrc.js' --fix 'src/**/*'",
     "test": "bash ops/entry.sh",
-    "load-test": "ts-node ./src/load/index.ts | pino-pretty -t"
+    "load-test-cyclical": "ts-node ./src/load/start-cyclical.ts | pino-pretty -t",
+    "load-test-concurrency": "ts-node ./src/load/start-concurrency.ts | pino-pretty -t"
   },
   "author": "",
   "license": "ISC",
