+++ modules/test-runner/src/load/start-concurrency.ts
@@ -1,0 +1,6 @@
+import { startServer } from "./setupServer";
+import { concurrencyTest } from "./test";
+
+startServer().then(async () => {
+  await concurrencyTest();
+});
