+++ modules/test-runner/src/load/start-cyclical.ts
@@ -1,0 +1,6 @@
+import { startServer } from "./setupServer";
+import { cyclicalTransferTest } from "./test";
+
+startServer().then(async () => {
+  await cyclicalTransferTest();
+});
