+++ modules/test-runner/src/load/test.ts
@@ -5,8 +5,7 @@ import PriorityQueue from "p-queue";
 import { env } from "../utils";
 
 import { AgentManager } from "./agent";
-
-import { carolEvts, logger } from "./index";
+import { carolEvts, logger } from "./setupServer";
 
 export const cyclicalTransferTest = async (): Promise<void> => {
   const agentService = await RestServerNodeService.connect(
