+++ ops/test-load.sh
@@ -8,7 +8,19 @@ project="`cat $root/package.json | grep '"name":' | head -n 1 | cut -d '"' -f 4`
 docker swarm init 2> /dev/null || true
 docker network create --attachable --driver overlay $project 2> /dev/null || true
 
-num_agents="${1:-5}"
+test_type="${1:cyclical}"
+num_agents="${2:-5}"
+if [[ "$test_type" == "cyclical" ]]
+then
+  echo "Running cyclical test"
+  test_cmd = "npm run load-test-cyclical"
+elif [[ "$test_type" == "concurrency" ]]
+  echo "Running concurrency test"
+  test_cmd = "npm run load-test-concurrency"
+else
+  echo "Unknown test type!"
+  exit 1
+fi
 
 sugar_daddy="candy maple cake sugar pudding cream honey rich smooth crumble sweet treat"
 
@@ -55,5 +67,5 @@ else
     $common \
     --entrypoint=bash \
     --volume="$root:/root" \
-    ${project}_builder -c "cd modules/test-runner && npm run load-test"
+    ${project}_builder -c "cd modules/test-runner && $test_cmd"
 fi
